Modernize SiStripConnectivity
- move contents of header to source
- use ESGetToken and  make_unique
