+++ CalibTracker/SiStripESProducers/plugins/geom/SiStripConnectivity.cc
@@ -1,13 +1,41 @@
-#include "CalibTracker/SiStripESProducers/plugins/geom/SiStripConnectivity.h"
-//#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "CalibTracker/Records/interface/SiStripFecCablingRcd.h"
+#include "CalibTracker/Records/interface/SiStripDetCablingRcd.h"
+#include "CalibFormats/SiStripObjects/interface/SiStripFecCabling.h"
+#include "CalibFormats/SiStripObjects/interface/SiStripDetCabling.h"
+
 #include <memory>
 
-SiStripConnectivity::SiStripConnectivity(const edm::ParameterSet& p) {
-  //the following lines are needed to tell the framework what data is being produced
-  setWhatProduced(this, &SiStripConnectivity::produceFecCabling);
-  setWhatProduced(this, &SiStripConnectivity::produceDetCabling);
-}
+class SiStripConnectivity: public edm::ESProducer {
+
+ public:
+
+  SiStripConnectivity( const edm::ParameterSet& );
+  ~SiStripConnectivity() override;
+
+  std::unique_ptr<SiStripFecCabling> produceFecCabling( const SiStripFecCablingRcd& );
+  std::unique_ptr<SiStripDetCabling> produceDetCabling( const SiStripDetCablingRcd& );
+
+ private:
+  struct FecTokens {
+    FecTokens(edm::ESConsumesCollector&& cc) { cc.setConsumes(fedCabling); }
+    edm::ESGetToken<SiStripFedCabling, SiStripFedCablingRcd> fedCabling;
+  };
+  struct DetTokens {
+    DetTokens(edm::ESConsumesCollector&& cc) { cc.setConsumes(fedCabling).setConsumes(tTopo); }
+    edm::ESGetToken<SiStripFedCabling, SiStripFedCablingRcd> fedCabling;
+    edm::ESGetToken<TrackerTopology, TrackerTopologyRcd> tTopo;
+  };
+  const FecTokens fecTokens_;
+  const DetTokens detTokens_;
+};
+
+SiStripConnectivity::SiStripConnectivity(const edm::ParameterSet& p)
+  : fecTokens_(setWhatProduced(this, &SiStripConnectivity::produceFecCabling)),
+    detTokens_(setWhatProduced(this, &SiStripConnectivity::produceDetCabling))
+{}
 
 SiStripConnectivity::~SiStripConnectivity() {
    // do anything here that needs to be done at desctruction time
@@ -17,23 +45,15 @@ SiStripConnectivity::~SiStripConnectivity() {
 // ------------ methods called to produce the data  ------------
 
 std::unique_ptr<SiStripFecCabling> SiStripConnectivity::produceFecCabling( const SiStripFecCablingRcd& iRecord ){
-  edm::ESHandle<SiStripFedCabling> pDD;
-  iRecord.getRecord<SiStripFedCablingRcd>().get(pDD );
   //here build an object of type SiStripFecCabling using  **ONLY** the information from class SiStripFedCabling, 
-  SiStripFecCabling * FecConnections = new SiStripFecCabling( *(pDD.product()));
-  return std::unique_ptr<SiStripFecCabling>( FecConnections );
+  return std::make_unique<SiStripFecCabling>(iRecord.get(fecTokens_.fedCabling));
 }
 
 std::unique_ptr<SiStripDetCabling> SiStripConnectivity::produceDetCabling( const SiStripDetCablingRcd& iRecord ){
-  edm::ESHandle<SiStripFedCabling> pDD;
-  iRecord.getRecord<SiStripFedCablingRcd>().get(pDD );
-  edm::ESHandle<TrackerTopology> tTopoHandle;
-  iRecord.getRecord<TrackerTopologyRcd>().get(tTopoHandle);
-  const TrackerTopology* const tTopo = tTopoHandle.product();
   //here build an object of type SiStripDetCabling using  **ONLY** the information from class SiStripFedCabling, 
-  SiStripDetCabling * DetConnections = new SiStripDetCabling( *(pDD.product()),tTopo);
-  return std::unique_ptr<SiStripDetCabling>( DetConnections );
+  return std::make_unique<SiStripDetCabling>(iRecord.get(detTokens_.fedCabling),
+                                             &iRecord.get(detTokens_.tTopo));
 }
 
-//define this as a plug-in
-// DEFINE_FWK_EVENTSETUP_MODULE(SiStripConnectivity);
+// define this as a plug-in
+DEFINE_FWK_EVENTSETUP_MODULE(SiStripConnectivity);
