+++ CalibTracker/SiStripESProducers/plugins/geom/modules.cc
@@ -7,9 +7,6 @@
 #include "CalibTracker/SiStripESProducers/plugins/geom/SiStripHashedDetIdESModule.h"
 DEFINE_FWK_EVENTSETUP_MODULE(SiStripHashedDetIdESModule);
 
-#include "CalibTracker/SiStripESProducers/plugins/geom/SiStripConnectivity.h"
-DEFINE_FWK_EVENTSETUP_MODULE(SiStripConnectivity);
-
 #include "CalibTracker/SiStripESProducers/plugins/geom/SiStripRegionConnectivity.h"
 DEFINE_FWK_EVENTSETUP_MODULE(SiStripRegionConnectivity);
